pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.265",
    "black>=22.12.0",
    "commitizen>=2.39.1",
    "pre-commit>=2.21.0",
    "pytest>=7.2.0",
    "allure-pytest>=2.12.0",
]

[tool.pylint."messages control"]
disable = ["missing-module-docstring"]
[tool.pylint.'MASTER']
ignore-patterns = ["^test.+"]


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version = "0.1.0"
update_changelog_on_bump = true
major_version_zero = true
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "MarlonOu", email = "jj80868588@gmail.com"},
]
dependencies = [
    "fastapi>=0.98.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=2.0.17",
    "mysql-connector-python>=8.0.33",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
]
requires-python = ">=3.11"
license = {text = "MIT"}
